Move query logging into clients
The old code had a lot of `print()` statements, to output the contents
of each DNS query & response.  That worked, but it was multi-line, and
it didn't use `logging`.

This moves all of that code into the `clients` module, logs DNS
queries/responses on a single line (in semi-readable format!), and does
away with `print()`.