Look up zone name from target name
In addition to the TARGET_NAME, the program needed the DOMAIN_NAME to be
specified, so that we could make our UPDATE messages relative to the
domain.  This changes the program to work out the domain name itself, by
doing a SOA lookup of the provided name.  The SOA-lookup function is
added to the existing resolver code.