Add support for Kerberos Credential Store Extensions
Kerberos authentication requires a couple of things:

* A Kerberos principal to get credentials for.  This can be an invidiaul
  (`somebody@stanford.edu`, for example, authenticated using a password)
  or a host/service (`host/blargh.stanford.edu@stanford.edu`, for
  example, authenticated using a keytab).

* A place to store credentials (a "credentials cache").

Normally, a separate program (such as `kinit` or `k5start`) is used to
get a Kerberos credential (a "ticket-granting ticket") for a
principal.  And, programs share a common credentials cache, which is
specified via the `KRB5CCNAME` environment variable.

With the Kerberos Credential Store Extension, you can specify a
credentials cache to use that is separate from the one set in the
environment.  In addition, you can specify a Keytab, removing the need
for a separate `kinit` or `k5start` to obtain a Kerberos credential.

The Kerberos Credential Store Extension is something that must be
supported by the underlying Keberos software, and by the Python GSSAPI
package.