Start adding support to clean up extra TXT records
The proper process for handling ACME DNS-01 Challenge records involves
deleting a TXT once the ACME server has seen it.  They are not supposed
to remain in the DNS after the challenge process is complete.

There may be ACME Challenge TXT records left behind, for various
reasons.  This adds an option to remove all TXT records found under
`_acme-challenge.{name}` before adding our challenge record to the DNS.

This can be dangerous: Other DNS-01 challenges may be in progress for
the name, and TXT records under `_acme-challenge.{name}` may be present
for other reasons.  So, to ensure the client means it, and understands
the risks, we will require that they present a unique code, derived from
the name.

This commit introduces the cleanup option, and does the cleanup code
validation.  TODO: Actual cleanup of records.