resolver: Raise exception on CNAME/DNAME
When doing a TXT or SOA lookup–especially when doing a SOA lookup–we
need to be on the lookout for CNAME and DNAME records.  So, update the
TXT- and SOA-lookup methods to check for CNAME/DNAME records in the
answer, and then raise an exception to the caller.  Also, provide a new
parameter, giving the client the option of ignoring the CNAME/DNAME.

Ignoring a CNAME/DNAME on a TXT lookup doesn't do anything different:
The returned TXT records are passed up to the client.

For SOA lookup, encountering a CNAME/DNAME triggers a fresh lookup, of
the original name's parent.  So, for example, if `blargh.stanford.edu`
returns a CNAME, and we're told to ignore that, then we'll do a lookup
of `stanford.edu` and return that.