Implement old-challenge cleanup
A previous commit started the work to clean up old challenges, but the
actual cleanup work wasn't implemented yet.  Until now!

The code does a lookup of TXT records for the ACME Challenge name, and
then issues separate UPDATEs to delete each record.  Since old
challenges do not matter to us, we do not wait for them to propagate, we
just continue on with inserting the challenge record that we care about.