resolver: Add tests for YXDOMAIN responses
With our test `localdomain.` zone, we have this record:

```
d  IN  DNAME  1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.1234567890.example.com.
```

With that DNAME, and a long-enough lookup, our test resolver will be
able to generate a YXDOMAIN response.

Add code to prepare a local resolver, and run YXDOMAIN tests.  Only do
all this if the PyTest "dns" mark is set.