Add check that the record is public
This adds code to check (via our non-caching resolver) to see if our
challenge record has made it into the public DNS.  We wait 1 minute
initially; after the check, if it fails, we enter a wait-then-check
loop.

TODO: Allow customization of the wait time.

Also, instead of encoding the challenge to bytes at the time of the
update, we do so immediately.  That makes the check a bytes-to-bytes
check.