Start adding certbot-specific code
This creates a set of classes for certbot to use.  Its structure is
similar to the CLI code, with a base class for common stuff, and
separate classes for the Stanford-specific and generic bits.

This commit gives enough code that certbot can be run, but the plugins
do not actually do anything yet.  Any attempt to actually use the
plugin will cause authentication to fail, because we're not actually
making changes to DNS yet!