Move waiting code into its own module
This moves the waiting code out of the CLI module, so it can be used
elsewhere.  It also turns it into a proper, type-safe set of classes.

We have an abstract base class, defining the interface to Waiter
objects.  We define a simple "Wait X seconds" class, and a more-complex
class customized to Stanford's DNS cycle.

A set of test cases is also provided.