bind/zone: Add TXT records to blargh
In order to test cleanup of old challenges, we need to make some TXT
records.

TXT records take the form of one or more strings, with no specific
encoding.  In Python-land, this is expressed as a tuple of bytestrings.
To ensure we have enough data to test, a single-entry and a multi-entry
record are created.