Add a module to handle cleanup
This is a module which handles the work around cleaning up of old
challenges.  It's intended to centralize code that's currently in
`sudns01.cli`, and which would've had to be copied into
`sudns01.certbot`.

The code handles generating cleanup challenges, checking challenge
validity, getting old challenges (via a DNS lookup), and creating the
nsupdate delete message that is needed to delete a TXT record.

The only thing this module doesn't do is to actually send out the
challenge messages, though that could be added in the future (though
testing it would be annoying).

Some changes were made during refactoring:

* DNS names are now brought in as `dns.name.Name` instances, not as
  strings.

* USER-FACING: As a result of the above change, DNS names now end in a
  `.`, so the cleanup challenge code is different.

One more note: Most methods allow specifying an ACME challenge domain
name as a parameter.  If not provided, one is automatically derived.
This is because certbot's code does the derivation for you.

Finally, also included is test code for all of the above.