test/resolver: Change test from 192.0.2.1 to 192.0.0.8
Some of the resolver tests need an IP address that does not respond.
So, I need an IP that is guaranteed to not have a DNS server.

RFC 5737 defines networks 192.0.2.0/24 (TEST-NET-1), 198.51.100.0/24
(TEST-NET-2), and 203.0.113.0/24 (TEST-NET-3) as "for use in
documentation."  The RFC goes on to say "These blocks are not for local
use."

Unfortunately, Cisco uses 192.0.2.1 as an IP for Cisco Wireless LAN
Controller (WLC) devices.  I ran into this, when I discovered 192.0.2.1
answering DNS queries in a Coupa Cafe.

So, a different test IP is needed.

192.0.0.8 is allocated in RFC 7600 §6.  §4.8 says that the IP is to be
used as an origin IP for certain ICMPv4 messages when used with IPv4
Residual Deployment ("4rd") in IPv6 networks.  As such, there's a good
chance this IP will never answer DNS queries.