challenge: Make acme_name required
Instead of making acme_name an optional class parameter, make it
required.

This vastly simplifies code, by removing the optional
`acme_challenge_name` parameter from many of our methods.

For calling code that does not know the ACME challenge name, a new
static method is created, which provides the ACME challenge name for a
given domain name.

Our post-init code is also updated, to check that acme_name is a
subdomain of the given domain.

The CLI code is also updated, to use the new parameter setup.