challenge/is_challenge_in_dns: Skip all tuples
In is_challenge_in_dns, when we encounter a tuple, we know that it does
not contain our record of interest.  That's because the resolver's
`get_txt` method will take single-item tuples and return just the one
item (instead of returning the entire tuple).  Since we know challenge
TXT records are single-item records, we can skip any tuples we find.