Add a GitHub Actions test workflow
This test workflow runs on every push and on every PR.  On Ubuntu &
Windows, it installs MIT Kerberos & Python.  On Ubuntu, it also installs
Bind.

On Ubuntu, all the PyTest tests are run.  Right now, no tests are run on
Windows, as no Makefile is available right now.

Tests are run against Python 3.10 through 3.13.

In addition, NOCOMMIT is checked for, and CodeQL is run.