Add a Makefile
This top-level Makefile is used to manage the sub-makefiles (for Bind
and MIT Kerberos), and to trigger execution of PyTest.  Targets are also
set up to handle doing an editable installation, installing test
dependencies, and starting/stopping the Test DNS & KDC.

Running `make` by itself lists the targets that most folks will be
interested in.