cerbot: Add code to make it usable!
This adds code to certbot, in order to make it usable.

First, a new abstract property, `_waiter`, is made.  Our subclasses
implement it, to provide an appropriate waiter.

The rest of the changes are made in the base class: We use the new
Challenge class to handle the creation of challenges, and all necessary
cleanup work.